Make base class for snippets and recording
